Scrape data in mod

+++ BetterTooltips/Common/Stats/WingStat.cs
@@ -1,10 +1,10 @@
 namespace BetterTooltips.Common.Stats
 {
-  public class WingStat(float flightTime, int height, float speedBonus)
+  public class WingStat(float flightTime, int height, int speedBonus)
   {
     public float FlightTime = flightTime;
     public int Height = height;
-    public float SpeedBonus = speedBonus;
-    public static WingStat Empty() => new(0.0f, 0, 0.0f);
+    public int SpeedBonus = speedBonus;
+    public static WingStat Empty() => new(0.0f, 0, 0);
   }
 }
